server/requirements.txt
APScheduler==3.10.1
asgiref==3.7.1
async-generator==1.10
attrs==23.1.0
autopep8==2.0.1
bcrypt==4.0.1
bson==0.5.10
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-py==113.0.5672.63
click==8.1.3
cryptography==40.0.2
cssselect==1.2.0
Django==4.2.1
dnspython==2.2.1
exceptiongroup==1.1.1
flake8==6.0.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
mccabe==0.7.0
Naked==0.1.32
nulltype==2.3.1
oauthlib==3.2.2
outcome==1.2.0
paypalrestsdk==1.13.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pymongo==4.3.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.9.1
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.16
venmo-api==0.3.1
waitress==2.1.2
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
